trimFactor is i32

+++ src/mongocxx/include/mongocxx/v_noabi/mongocxx/options/range.hpp
@@ -65,10 +65,10 @@ class range {
     const stdx::optional<std::int64_t>& sparsity() const;
 
     /// @brief Sets `RangeOpts.trimFactor`.
-    range& trimFactor(std::int64_t value);
+    range& trimFactor(std::int32_t value);
 
     /// @brief Gets `RangeOpts.trimFactor`.
-    const stdx::optional<std::int64_t>& trim_factor() const;
+    const stdx::optional<std::int32_t>& trim_factor() const;
 
     /// @brief Sets `RangeOpts.precision`.
     /// @note May only be set for `double` or `decimal128`.
@@ -82,7 +82,7 @@ class range {
     stdx::optional<bsoncxx::v_noabi::types::bson_value::view_or_value> _min;
     stdx::optional<bsoncxx::v_noabi::types::bson_value::view_or_value> _max;
     stdx::optional<std::int64_t> _sparsity;
-    stdx::optional<std::int64_t> _trim_factor;
+    stdx::optional<std::int32_t> _trim_factor;
     stdx::optional<std::int32_t> _precision;
 };
