+++ src/mongocxx/lib/mongocxx/v_noabi/mongocxx/options/range.cpp
@@ -43,7 +43,7 @@ range& range::sparsity(std::int64_t value) {
     return *this;
 }
 
-range& range::trim_factor(std::int64_t value) {
+range& range::trim_factor(std::int32_t value) {
     _trim_factor = value;
     return *this;
 }
@@ -52,7 +52,7 @@ const stdx::optional<std::int64_t>& range::sparsity() const {
     return _sparsity;
 }
 
-const stdx::optional<std::int64_t>& range::trim_factor() const {
+const stdx::optional<std::int32_t>& range::trim_factor() const {
     return _trim_factor;
 }
